Clean up CHANGELOG and README Files
Clean up the CHANGELOG and the README files. The CHANGELOG got cleaned
in its entirety, and only a new "Unreleased" version is added with the
comment that also is present in the other projects (splitting the repo).

The README was pruned from non-relevant vscode-psl related sections, and
the history was added, pointing to the split commit.