Fix Build and Test
Fix the build and test workflow, as the build npm script does not exist
(yet). Update the test script to not exit on error because there are no
tests yet (added `|| exit 0`). This does need to be removed once tests
are introduces, as now the test script will always pass.