Add test and build related changes
This commit introduces a simple unit test to complete the full set of
scripts that are used to build and publish this package. It is
simplified from the psl-parser and psl-lint projects, as this package is
using the builtin test framework of node, and does not required two
diferent tsconfig files. The other projects might follow soon.